+++ packages/mobile/src/fiatExchanges/ProviderOptionsScreen.test.tsx
@@ -5,8 +5,7 @@ import { Text } from 'react-native'
 import { fireEvent, render, waitForElement } from 'react-native-testing-library'
 import { Provider } from 'react-redux'
 import { PaymentMethod } from 'src/fiatExchanges/FiatExchangeOptions'
-import ProviderOptionsScreen from 'src/fiatExchanges/ProviderOptionsScreen'
-import { CicoProvider } from 'src/fiatExchanges/utils'
+import ProviderOptionsScreen, { CicoProvider } from 'src/fiatExchanges/ProviderOptionsScreen'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
